Include import-validator as a hidden algorand-indexer subcommand.
